future-self/backend/requirements.txt
# FastAPI Backend Dependencies for Future Self App
# Core dependencies with specific versions below

# For astrology integration
pyswisseph # Swiss Ephemeris for precise astrological calculations
pytz # For timezone handling in astrology

# After installing spacy, you'll need to download a model, e.g.:
# python -m spacy download en_core_web_sm

# Core FastAPI framework
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for API requests
requests==2.31.0

# Supabase integration
supabase==2.3.4
postgrest==0.13.2

# Audio processing dependencies
whisper==1.1.10
TTS==0.22.0
pyogg==0.6.14a1

# Audio file handling (built-in modules don't need to be listed)
# wave, struct, base64, tempfile, shutil, subprocess are built-in

# Data handling
pydantic==2.5.0

# CORS middleware (included in fastapi[all])

# Environment variables
python-dotenv==1.0.0

# Additional audio processing (if needed)
# ffmpeg-python==0.2.0  # Alternative to subprocess ffmpeg calls
librosa==0.10.1       # Advanced audio analysis
soundfile==0.12.1     # Audio file I/O

# NLP and Machine Learning dependencies
transformers==4.36.0  # Hugging Face transformers for emotion detection
torch==2.1.0          # PyTorch for ML models
numpy==1.24.3         # Numerical computing
scipy==1.11.4         # Scientific computing
scikit-learn==1.3.2   # Machine learning utilities
spacy==3.7.2          # Natural language processing
textblob==0.17.1      # Simple text processing
vaderSentiment==3.3.2 # Sentiment analysis
detoxify==0.5.2       # Toxicity detection
langdetect==1.0.9     # Language detection
redis==5.0.1          # For caching and analytics
pandas>=1.4,<2.0      # Data manipulation (compatible with TTS)
matplotlib==3.8.2     # For analytics visualization
seaborn==0.13.0       # Statistical data visualization